sources/inbox/auto-apply-pro/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.9.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
alembic==1.13.2
typer==0.12.5
playwright==1.48.0
httpx==0.27.2
beautifulsoup4==4.12.3
python-docx==1.1.2
reportlab==4.2.5
pandas==2.2.2
pytest==8.3.2
pytest-playwright==0.5.7
